requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
alembic==1.16.5
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.1
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.49.1
google-genai==1.68.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.1.1
kaggle==2.0.0
kagglehub==0.3.13
kagglesdk==0.1.16
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
Mako==1.3.10
MarkupSafe==3.0.3
marshmallow==3.26.2
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.1.3
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==2.2.3
propcache==0.4.1
protobuf==7.34.1
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg-pool==3.3.0
psycopg2==2.9.11
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
RapidFuzz==3.14.3
requests==2.33.0
requests-toolbelt==1.0.0
scikit-learn==1.8.0
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==1.0.0
tenacity==9.1.4
text-unidecode==1.3
threadpoolctl==3.6.0
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchfiles==1.1.1
webencodings==0.5.1
websockets==16.0
xgboost-cpu==3.2.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
